Refactor inventory component for better modularity
