+++ client/src/app/inventory/page.tsx
@@ -1,926 +1,360 @@
 "use client";
 
 import React, { useState, useMemo, useCallback } from "react";
-import { DataGrid, GridColDef, GridRenderCellParams } from "@mui/x-data-grid";
+import { DataGrid } from "@mui/x-data-grid";
 import {
- Box,
- Button,
- Modal,
- TextField,
- Typography,
- Alert,
- List,
- ListItem,
- Divider,
- IconButton,
- Chip,
- Tooltip,
- Paper,
- Rating as MuiRating,
- ToggleButton,
- ToggleButtonGroup,
+  Box,
+  Button,
+  TextField,
+  Typography,
+  ToggleButton,
+  ToggleButtonGroup,
 } from "@mui/material";
 import {
- Product,
- useGetProductsQuery,
- useUpdateProductFieldsMutation,
+  Product,
+  useGetProductsQuery,
+  useUpdateProductFieldsMutation,
 } from "@/state/api";
 import { GridToolbar } from "@mui/x-data-grid";
 import { dataGridStyles } from "@/styles/dataGridStyles";
-import Image from "next/image";
 import { useAppSelector } from "@/app/redux";
-import {
- CheckCircle,
- Add,
- Remove,
- Visibility,
- Star,
- StarBorder,
-} from "@mui/icons-material";
+
+// Import custom components and hooks
+import RestockModal from "./RestockModal";
+import { getGridColumns } from "./GridColumns";
+import useProductUnits from "./useProductUnits";
+import { SelectedProductInfo } from "../types/productTypes";
+import ProductDetailModal from "./ProductDetailModal";
 
 const Inventory = () => {
- // Get dark mode state from Redux
- const isDarkMode = useAppSelector((state) => state.global.isDarkMode);
+  // Get dark mode state from Redux
+  const isDarkMode = useAppSelector((state) => state.global.isDarkMode);
 
- // State declarations
- const [searchQuery, setSearchQuery] = useState<string>("");
- const [selectedProductIds, setSelectedProductIds] = useState<string[]>([]);
- const [openRestockModal, setOpenRestockModal] = useState<boolean>(false);
- const [viewProductModal, setViewProductModal] = useState<boolean>(false);
- const [selectedViewProduct, setSelectedViewProduct] = useState<Product | null>(
-  null
- );
- const [featuredProducts, setFeaturedProducts] = useState<Set<string>>(
-  new Set()
- );
+  // State declarations
+  const [searchQuery, setSearchQuery] = useState<string>("");
+  const [selectedProductIds, setSelectedProductIds] = useState<string[]>([]);
+  const [openRestockModal, setOpenRestockModal] = useState<boolean>(false);
+  const [viewProductModal, setViewProductModal] = useState<boolean>(false);
+  const [selectedViewProduct, setSelectedViewProduct] = useState<Product | null>(null);
+  const [featuredProducts, setFeaturedProducts] = useState<Set<string>>(new Set());
 
- // Filter mode: "all", "featured", "regular"
- const [filterMode, setFilterMode] = useState<string>("all");
+  // Filter mode: "all", "featured", "regular"
+  const [filterMode, setFilterMode] = useState<string>("all");
 
- // State for individual product restock quantities
- const [productRestockQuantities, setProductRestockQuantities] = useState<{
-  [key: string]: number;
- }>({});
- const [activeProductId, setActiveProductId] = useState<string | null>(null);
+  // State for individual product restock quantities
+  const [productRestockQuantities, setProductRestockQuantities] = useState<{
+    [key: string]: number;
+  }>({});
+  const [activeProductId, setActiveProductId] = useState<string | null>(null);
 
- // API hooks
- const { data: products = [], isLoading, refetch } = useGetProductsQuery();
- const [updateProductFields] = useUpdateProductFieldsMutation();
+  // Custom hook for product units
+  const { getProductUnit } = useProductUnits();
 
- 
-// Replace your current toggleFeatured function with this:
-const toggleFeatured = useCallback((productId: string) => {
-  const newFeaturedProducts = new Set(featuredProducts);
-  if (newFeaturedProducts.has(productId)) {
-    newFeaturedProducts.delete(productId);
-  } else {
-    newFeaturedProducts.add(productId);
-  }
-  setFeaturedProducts(newFeaturedProducts);
-}, [featuredProducts]);
+  // API hooks
+  const { data: products = [], isLoading, refetch } = useGetProductsQuery();
+  const [updateProductFields] = useUpdateProductFieldsMutation();
+  
+  // View product handlers
+  const handleViewProduct = useCallback((product: Product) => {
+    setSelectedViewProduct(product);
+    setViewProductModal(true);
+  }, []);
 
- // Function to get display list of selected products
- const getSelectedProductsInfo = () => {
-  if (!selectedProductIds.length) return [];
+  const toggleFeatured = useCallback((productId: string) => {
+    const newFeaturedProducts = new Set(featuredProducts);
+    if (newFeaturedProducts.has(productId)) {
+      newFeaturedProducts.delete(productId);
+    } else {
+      newFeaturedProducts.add(productId);
+    }
+    setFeaturedProducts(newFeaturedProducts);
+  }, [featuredProducts]);
+  // Get DataGrid columns configuration
+  const columns = useMemo(() => getGridColumns({
+    isDarkMode,
+    featuredProducts,
+    toggleFeatured,
+    getProductUnit,
+    handleViewProduct,
+  }), [isDarkMode, featuredProducts, toggleFeatured, getProductUnit, handleViewProduct]);
 
-  return selectedProductIds.map((id) => {
-   const product = products.find((p) => p.productId === id);
-   return {
-    id: product?.productId || "",
-    name: product?.name || "Unknown product",
-    currentStock: product?.stockQuantity || 0,
-    // Get restock quantity from state or default to 0
-    restockQuantity: productRestockQuantities[id] || 0,
-   };
-  });
- };
+  // Toggle featured status for a product
 
- // Modal handlers
- const handleOpenRestockModal = () => {
-  // Initialize the first selected product as active
-  if (selectedProductIds.length > 0) {
-   setActiveProductId(selectedProductIds[0]);
-  }
 
-  // Reset all restock quantities to 0
-  const initialQuantities = selectedProductIds.reduce((acc, id) => {
-   acc[id] = 0;
-   return acc;
-  }, {} as { [key: string]: number });
+  // Function to get display list of selected products
+  const getSelectedProductsInfo = useCallback((): SelectedProductInfo[] => {
+    if (!selectedProductIds.length) return [];
 
-  setProductRestockQuantities(initialQuantities);
-  setOpenRestockModal(true);
- };
+    return selectedProductIds.map((id) => {
+      const product = products.find((p) => p.productId === id);
+      return {
+        id: product?.productId || "",
+        name: product?.name || "Unknown product",
+        currentStock: product?.stockQuantity || 0,
+        // Get restock quantity from state or default to 0
+        restockQuantity: productRestockQuantities[id] || 0,
+        // Add unit information
+        unit: product ? getProductUnit(product) : "Units",
+      };
+    });
+  }, [selectedProductIds, products, productRestockQuantities, getProductUnit]);
 
- const handleCloseRestockModal = () => {
-  setOpenRestockModal(false);
-  setActiveProductId(null);
-  setProductRestockQuantities({});
- };
+  // Modal handlers
+  const handleOpenRestockModal = useCallback(() => {
+    // Initialize the first selected product as active
+    if (selectedProductIds.length > 0) {
+      setActiveProductId(selectedProductIds[0]);
+    }
 
- // View product handlers
- const handleViewProduct = (product: Product) => {
-  setSelectedViewProduct(product);
-  setViewProductModal(true);
- };
+    // Reset all restock quantities to 0
+    const initialQuantities = selectedProductIds.reduce((acc, id) => {
+      acc[id] = 0;
+      return acc;
+    }, {} as { [key: string]: number });
 
- const handleCloseViewProduct = () => {
-  setViewProductModal(false);
-  setSelectedViewProduct(null);
- };
+    setProductRestockQuantities(initialQuantities);
+    setOpenRestockModal(true);
+  }, [selectedProductIds]);
 
- // Handle quantity change for a specific product
- const handleQuantityChange = (id: string, value: number) => {
-  if (value >= 0) {
-   setProductRestockQuantities((prev) => ({
-    ...prev,
-    [id]: value,
-   }));
-  }
- };
+  const handleCloseRestockModal = useCallback(() => {
+    setOpenRestockModal(false);
+    setActiveProductId(null);
+    setProductRestockQuantities({});
+  }, []);
 
- // Set active product
- const handleSetActiveProduct = (id: string) => {
-  setActiveProductId(id);
- };
 
- // Increment/decrement quantity
- const incrementQuantity = (id: string) => {
-  const currentQty = productRestockQuantities[id] || 0;
-  handleQuantityChange(id, currentQty + 1);
- };
 
- const decrementQuantity = (id: string) => {
-  const currentQty = productRestockQuantities[id] || 0;
-  if (currentQty > 0) {
-   handleQuantityChange(id, currentQty - 1);
-  }
- };
+  const handleCloseViewProduct = useCallback(() => {
+    setViewProductModal(false);
+    setSelectedViewProduct(null);
+  }, []);
 
- // Handle filter change
- const handleFilterChange = (
-  event: React.MouseEvent<HTMLElement>,
-  newFilterMode: string
- ) => {
-  if (newFilterMode !== null) {
-   setFilterMode(newFilterMode);
-  }
- };
+  // Handle quantity change for a specific product
+  const handleQuantityChange = useCallback((id: string, value: number) => {
+    if (value >= 0) {
+      setProductRestockQuantities((prev) => ({
+        ...prev,
+        [id]: value,
+      }));
+    }
+  }, []);
 
- // Restock handler with individual quantities and auto-clear selection
- const handleRestockProducts = async () => {
-  try {
-   console.group("Restock Process");
+  // Set active product
+  const handleSetActiveProduct = useCallback((id: string) => {
+    setActiveProductId(id);
+  }, []);
 
-   // Filter out products with zero restock quantity
-   const productsToUpdate = Object.entries(productRestockQuantities)
-    .filter(([_, qty]) => qty > 0)
-    .map(([id, qty]) => ({ id, qty }));
+  // Increment/decrement quantity
+  const incrementQuantity = useCallback((id: string) => {
+    const currentQty = productRestockQuantities[id] || 0;
+    handleQuantityChange(id, currentQty + 1);
+  }, [productRestockQuantities, handleQuantityChange]);
 
-   if (productsToUpdate.length === 0) {
-    console.log("No products to update");
-    return;
-   }
+  const decrementQuantity = useCallback((id: string) => {
+    const currentQty = productRestockQuantities[id] || 0;
+    if (currentQty > 0) {
+      handleQuantityChange(id, currentQty - 1);
+    }
+  }, [productRestockQuantities, handleQuantityChange]);
 
-   await Promise.all(
-    productsToUpdate.map(async ({ id: productId, qty }) => {
-     const product = products.find((p) => p.productId === productId);
-     if (!product) {
-      console.warn(`Product ${productId} not found`);
-      return;
-     }
+  // Handle filter change
+  const handleFilterChange = useCallback((
+    event: React.MouseEvent<HTMLElement>,
+    newFilterMode: string
+  ) => {
+    if (newFilterMode !== null) {
+      setFilterMode(newFilterMode);
+    }
+  }, []);
 
-     const newStock = product.stockQuantity + qty;
+  // Restock handler with individual quantities and auto-clear selection
+  const handleRestockProducts = useCallback(async () => {
+    try {
+      console.group("Restock Process");
 
-     console.log("Updating product:", {
-      id: productId,
-      name: product.name,
-      from: product.stockQuantity,
-      to: newStock,
-      addedQuantity: qty,
-     });
+      // Filter out products with zero restock quantity
+      const productsToUpdate = Object.entries(productRestockQuantities)
+        .filter(([_, qty]) => qty > 0)
+        .map(([id, qty]) => ({ id, qty }));
 
-     // Use the dedicated fields update mutation
-     await updateProductFields({
-      productId,
-      updateData: { stockQuantity: newStock },
-     }).unwrap();
-    })
-   );
+      if (productsToUpdate.length === 0) {
+        console.log("No products to update");
+        return;
+      }
 
-   await refetch();
-   console.log("Restock completed successfully");
+      await Promise.all(
+        productsToUpdate.map(async ({ id: productId, qty }) => {
+          const product = products.find((p) => p.productId === productId);
+          if (!product) {
+            console.warn(`Product ${productId} not found`);
+            return;
+          }
 
-   // Clear selection after successful restock
-   setSelectedProductIds([]);
-  } catch (error) {
-   console.error("Restock failed:", error);
-  } finally {
-   handleCloseRestockModal();
-   console.groupEnd();
-  }
- };
+          const newStock = product.stockQuantity + qty;
 
- // Check if any product has a restock quantity > 0
- const hasProductsToRestock = useMemo(() => {
-  return Object.values(productRestockQuantities).some((qty) => qty > 0);
- }, [productRestockQuantities]);
+          console.log("Updating product:", {
+            id: productId,
+            name: product.name,
+            from: product.stockQuantity,
+            to: newStock,
+            addedQuantity: qty,
+          });
 
- // Count products that will be restocked
- const productsToRestockCount = useMemo(() => {
-  return Object.values(productRestockQuantities).filter((qty) => qty > 0)
-   .length;
- }, [productRestockQuantities]);
-
- // Generate star rating component
- const renderStarRating = (rating: number = 0) => {
-  return (
-   <div className="flex justify-center">
-    <MuiRating
-     value={rating}
-     readOnly
-     precision={0.5}
-     size="small"
-     sx={{
-      color: isDarkMode ? "#FFD700" : "#FFA000",
-     }}
-    />
-   </div>
-  );
- };
-
- // DataGrid columns
- const columns: GridColDef[] = useMemo(
-  () => [
-   {
-    field: "featured",
-    headerName: "⭐",
-    width: 60,
-    sortable: true,
-    filterable: false,
-    renderCell: (params) => {
-     const isFeatured = featuredProducts.has(params.row.productId);
-     return (
-      <Box sx={{ display: "flex", justifyContent: "center", width: "100%" }}>
-       <IconButton
-        onClick={(e) => {
-         e.stopPropagation();
-         toggleFeatured(params.row.productId);
-        }}
-        size="small"
-        sx={{
-         color: isFeatured ? "gold" : "gray",
-         "&:hover": {
-          bgcolor: "rgba(255, 215, 0, 0.1)",
-         },
-        }}
-       >
-        {isFeatured ? (
-         <Star fontSize="small" />
-        ) : (
-         <StarBorder fontSize="small" />
-        )}
-       </IconButton>
-      </Box>
-     );
-    },
-    getValue: (params: { id: string | number; row: any }) => {
-      const id = params.id?.toString() || '';
-      return featuredProducts.has(id) ? 1 : 0;
-    },
-   },
-   {
-    field: "image",
-    headerName: "Image",
-    width: 100,
-    renderCell: (params) => (
-     <div
-      style={{
-       display: "flex",
-       justifyContent: "center",
-       alignItems: "center",
-       width: "100%",
-      }}
-     >
-      <Image
-       src={
-        params.row.image
-         ? `http://localhost:8000${params.row.image}`
-         : "/assets/default-image.png"
-       }
-       alt={params.row.name}
-       width={50}
-       height={50}
-       style={{ objectFit: "contain", borderRadius: "4px" }}
-       onError={(e) => {
-        (e.target as HTMLImageElement).src = "/assets/default-image.png";
-       }}
-      />
-     </div>
-    ),
-    sortable: false,
-    filterable: false,
-   },
-   { field: "productId", headerName: "ID", width: 120 },
-   { field: "name", headerName: "Name", flex: 0.5 },
-   {
-    field: "price",
-    headerName: "Price",
-    type: "number",
-    flex: 0.4,
-    headerAlign: "left",
-    valueGetter: (value, row) => `$${row.price}`,
-   },
-   // Category column
-   {
-    field: "category",
-    headerName: "Category",
-    flex: 0.4,
-    renderCell: (params) => {
-      const category = params.value || "Uncategorized";
-      
-      return (
-        <div style={{ display: "flex", justifyContent: "center", width: "100%" }}>
-          <Chip 
-            label={category} 
-            size="small"
-            variant="outlined"
-            sx={{ 
-              borderRadius: '4px',
-              bgcolor: isDarkMode ? 'rgba(0, 0, 0, 0.2)' : 'rgba(0, 0, 0, 0.05)',
-              borderColor: isDarkMode ? 'rgba(255, 255, 255, 0.12)' : 'rgba(0, 0, 0, 0.12)',
-            }}
-          />
-        </div>
+          // Use the dedicated fields update mutation
+          await updateProductFields({
+            productId,
+            updateData: { stockQuantity: newStock },
+          }).unwrap();
+        })
       );
-    },
-  },
-   {
-    field: "rating",
-    headerName: "Rating",
-    type: "number",
-    flex: 0.4,
-    renderCell: (params) => renderStarRating(params.value),
-   },
-   {
-    field: "stockQuantity",
-    headerName: "Stock Quantity",
-    flex: 0.4,
-    headerAlign: "right",
-    renderCell: (params) => {
-     // Determine style based on stock quantity
-     const isHighStock = params.value >= 100;
 
-     // Color values for light and dark mode
-     const colors = isDarkMode
-      ? {
-         highBg: "rgba(46, 204, 113, 0.25)",
-         highText: "#2ecc71",
-         lowBg: "rgba(231, 76, 60, 0.25)",
-         lowText: "#e74c3c",
-        }
-      : {
-         highBg: "rgba(46, 204, 113, 0.15)",
-         highText: "#27ae60",
-         lowBg: "rgba(231, 76, 60, 0.15)",
-         lowText: "#c0392b",
-        };
+      await refetch();
+      console.log("Restock completed successfully");
 
-     return (
-      <div style={{ display: "flex", justifyContent: "center", width: "100%" }}>
-       <span
-        style={{
-         display: "flex",
-         justifyContent: "center",
-         alignItems: "center",
-         backgroundColor: isHighStock ? colors.highBg : colors.lowBg,
-         color: isHighStock ? colors.highText : colors.lowText,
-         fontWeight: 600,
-         padding: "4px 12px",
-         borderRadius: "16px",
-         fontFamily:
-          '"SFMono-Regular", Consolas, "Liberation Mono", Menlo, monospace',
-         minWidth: "40px",
-         textAlign: "center",
-         height: "30px",
-        }}
-       >
-        {params.value}
-       </span>
-      </div>
-     );
-    },
-   },
-   {
-    field: "actions",
-    headerName: "Actions",
-    width: 100,
-    sortable: false,
-    filterable: false,
-    renderCell: (params: GridRenderCellParams) => (
-     <Box sx={{ display: "flex", justifyContent: "center", width: "100%" }}>
-      <Tooltip title="View Details">
-       <IconButton
-        onClick={(e) => {
-         e.stopPropagation();
-         handleViewProduct(params.row);
-        }}
-        size="small"
-        sx={{
-         color: "primary.main",
-         "&:hover": {
-          bgcolor: "rgba(40, 100, 150, 0.08)",
-         },
-        }}
-       >
-        <Visibility />
-       </IconButton>
-      </Tooltip>
-     </Box>
-    ),
-   },
-  ],
-  [isDarkMode, featuredProducts, renderStarRating, toggleFeatured ] // Add featuredProducts as a dependency
- );
+      // Clear selection after successful restock
+      setSelectedProductIds([]);
+    } catch (error) {
+      console.error("Restock failed:", error);
+    } finally {
+      handleCloseRestockModal();
+      console.groupEnd();
+    }
+  }, [productRestockQuantities, products, updateProductFields, refetch, handleCloseRestockModal]);
 
- // Filter products based on search query and featured status
- const filteredProducts = useMemo(() => {
-  let filtered = products.filter((product: Product) =>
-   product.name.toLowerCase().includes(searchQuery.toLowerCase())
+  // Check if any product has a restock quantity > 0
+  const hasProductsToRestock = useMemo(() => {
+    return Object.values(productRestockQuantities).some((qty) => qty > 0);
+  }, [productRestockQuantities]);
+
+  // Count products that will be restocked
+  const productsToRestockCount = useMemo(() => {
+    return Object.values(productRestockQuantities).filter((qty) => qty > 0).length;
+  }, [productRestockQuantities]);
+
+  // Filter products based on search query and featured status
+  const filteredProducts = useMemo(() => {
+    let filtered = products.filter((product: Product) =>
+      product.name.toLowerCase().includes(searchQuery.toLowerCase())
+    );
+
+    // Apply featured filter if needed
+    if (filterMode === "featured") {
+      filtered = filtered.filter((product) =>
+        featuredProducts.has(product.productId)
+      );
+    } else if (filterMode === "regular") {
+      filtered = filtered.filter(
+        (product) => !featuredProducts.has(product.productId)
+      );
+    }
+
+    return filtered;
+  }, [products, searchQuery, featuredProducts, filterMode]);
+
+  // Get selected products information for the modal
+  const selectedProducts = getSelectedProductsInfo();
+
+  // Count of featured products
+  const featuredCount = useMemo(
+    () => products.filter((p) => featuredProducts.has(p.productId)).length,
+    [products, featuredProducts]
   );
 
-  // Apply featured filter if needed
-  if (filterMode === "featured") {
-   filtered = filtered.filter((product) =>
-    featuredProducts.has(product.productId)
-   );
-  } else if (filterMode === "regular") {
-   filtered = filtered.filter(
-    (product) => !featuredProducts.has(product.productId)
-   );
-  }
-
-  return filtered;
- }, [products, searchQuery, featuredProducts, filterMode]);
-
- // Get selected products information for the modal
- const selectedProducts = getSelectedProductsInfo();
-
- // Count of featured products
- const featuredCount = useMemo(
-  () => products.filter((p) => featuredProducts.has(p.productId)).length,
-  [products, featuredProducts]
- );
-
- return (
-  <Box m="1.5rem 2.5rem">
-   <Box display="flex" flexDirection="column" gap={2}>
-    <Box className="flex justify-between items-center gap-4">
-     <TextField
-      label="Search Inventory"
-      variant="outlined"
-      value={searchQuery}
-      onChange={(e) => setSearchQuery(e.target.value)}
-     />
-     <Button
-      color="primary"
-      variant="contained"
-      onClick={handleOpenRestockModal}
-      disabled={selectedProductIds.length === 0}
-     >
-      Restock <span className="hidden md:block"> Selected Products</span>
-     </Button>
-    </Box>
-
-    {/* Featured filter toggle */}
-    <Box display="flex"  className="flex justify-between gap-3 items-center">
-     <ToggleButtonGroup
-     className="flex justify-between gap-3 items-center"
-      value={filterMode}
-      exclusive
-      onChange={handleFilterChange}
-      aria-label="product filter"
-      size="small"
-     >
-      <ToggleButton value="all" aria-label="all products">
-       <span className="text-[10px] font-medium md:text-base">All Products </span> 
-      </ToggleButton>
-      <ToggleButton value="featured" aria-label="featured products">
-      <span className="text-[10px] font-medium md:text-base">Featured ({featuredCount}) </span>  
-      </ToggleButton>
-      <ToggleButton value="regular" aria-label="regular products">
-      <span className="text-[10px] font-medium md:text-base">Regular </span> 
-      </ToggleButton>
-     </ToggleButtonGroup>
-
-     <Typography variant="body2" color="text.secondary">
-       <span className="hidden md:block rounded-full text-[10px] bg-green-100 text-green-800 md:text-base p-3">{filteredProducts.length} products found </span>
-     </Typography>
-    </Box>
-   </Box>
-
-   <Box
-    mt="20px"
-    height="75vh"
-    sx={{ "& .MuiDataGrid-root": { border: "none" } }}
-   >
-    <DataGrid
-     loading={isLoading}
-     rows={filteredProducts}
-     columns={columns}
-     getRowId={(row) => row.productId}
-     checkboxSelection
-     onRowSelectionModelChange={(newSelection) => {
-      setSelectedProductIds(newSelection as string[]);
-     }}
-     rowSelectionModel={selectedProductIds}
-     slots={{ toolbar: GridToolbar }}
-     sx={dataGridStyles}
-    />
-   </Box>
-
-   {/* Restock Modal */}
-   <Modal open={openRestockModal} onClose={handleCloseRestockModal}>
-    <Box
-     sx={{
-      position: "absolute",
-      top: "50%",
-      left: "50%",
-      transform: "translate(-50%, -50%)",
-      bgcolor: "background.paper",
-      boxShadow: 24,
-      p: 4,
-      minWidth: 400,
-      maxWidth: 600,
-      maxHeight: "90vh",
-      overflowY: "auto",
-      borderRadius: "8px",
-     }}
-    >
-     <Typography
-      variant="h5"
-      component="h2"
-      sx={{ color: "text.primary", mb: 2 }}
-     >
-      Restock Products
-     </Typography>
-
-     {selectedProductIds.length > 1 && (
-      <Alert severity="info" sx={{ mb: 3 }}>
-       You can set different restock quantities for each product.
-       {productsToRestockCount > 0 && (
-        <Box mt={1}>
-         <Chip
-          label={`${productsToRestockCount} product${
-           productsToRestockCount > 1 ? "s" : ""
-          } will be restocked`}
-          color="primary"
-          size="small"
-         />
-        </Box>
-       )}
-      </Alert>
-     )}
-
-     {selectedProducts.length > 0 && (
-      <Box sx={{ mb: 3 }}>
-       <Typography variant="subtitle1" sx={{ color: "text.primary", mb: 1 }}>
-        Selected Products:
-       </Typography>
-       <Box
-        sx={{
-         border: "1px solid",
-         borderColor: "divider",
-         borderRadius: "8px",
-         overflow: "hidden",
-        }}
-       >
-        <List disablePadding>
-         {selectedProducts.map((product, index) => {
-          const isActive = product.id === activeProductId;
-          const willRestock = product.restockQuantity > 0;
-
-          return (
-           <React.Fragment key={product.id}>
-            {index > 0 && <Divider />}
-            <ListItem
-             sx={{
-              p: 0,
-              cursor: "pointer",
-              bgcolor: isActive ? "rgba(25, 118, 210, 0.08)" : "transparent",
-              transition: "background-color 0.2s",
-             }}
-             onClick={() => handleSetActiveProduct(product.id)}
-            >
-             <Box
-              sx={{
-               width: "100%",
-               p: 2,
-               display: "flex",
-               flexDirection: "column",
-               gap: 1,
-              }}
-             >
-              <Box
-               sx={{
-                display: "flex",
-                justifyContent: "space-between",
-                alignItems: "center",
-               }}
-              >
-               <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-                <Typography
-                 variant="body1"
-                 sx={{
-                  color: "text.primary",
-                  fontWeight: isActive ? 600 : 400,
-                 }}
-                >
-                 {product.name}
-                </Typography>
-                {willRestock && (
-                 <Tooltip title="Will be restocked">
-                  <CheckCircle fontSize="small" color="success" />
-                 </Tooltip>
-                )}
-               </Box>
-               <Typography variant="body2" color="text.secondary">
-                Current: {product.currentStock}
-               </Typography>
-              </Box>
-
-              {isActive && (
-               <Box
-                sx={{
-                 display: "flex",
-                 alignItems: "center",
-                 gap: 1,
-                 mt: 1,
-                }}
-               >
-                <IconButton
-                 size="small"
-                 onClick={(e) => {
-                  e.stopPropagation();
-                  decrementQuantity(product.id);
-                 }}
-                 disabled={product.restockQuantity <= 0}
-                >
-                 <Remove fontSize="small" />
-                </IconButton>
-
-                <TextField
-                 value={product.restockQuantity}
-                 onChange={(e) => {
-                  const val = parseInt(e.target.value);
-                  handleQuantityChange(product.id, isNaN(val) ? 0 : val);
-                 }}
-                 type="number"
-                 size="small"
-                 inputProps={{ min: 0, style: { textAlign: "center" } }}
-                 sx={{ width: 80 }}
-                 variant="outlined"
-                 onClick={(e) => e.stopPropagation()}
-                />
-
-                <IconButton
-                 size="small"
-                 onClick={(e) => {
-                  e.stopPropagation();
-                  incrementQuantity(product.id);
-                 }}
-                >
-                 <Add fontSize="small" />
-                </IconButton>
-
-                {product.restockQuantity > 0 && (
-                 <Typography
-                  variant="body2"
-                  sx={{ ml: 1, color: "success.main" }}
-                 >
-                  New: {product.currentStock + product.restockQuantity}
-                 </Typography>
-                )}
-               </Box>
-              )}
-             </Box>
-            </ListItem>
-           </React.Fragment>
-          );
-         })}
-        </List>
-       </Box>
-      </Box>
-     )}
-
-     <Box sx={{ display: "flex", justifyContent: "space-between", mt: 3 }}>
-      <Button variant="outlined" onClick={handleCloseRestockModal}>
-       Cancel
-      </Button>
-      <Button
-       variant="contained"
-       color="primary"
-       onClick={handleRestockProducts}
-       disabled={!hasProductsToRestock}
-      >
-       {productsToRestockCount === 0
-        ? "Restock Products"
-        : `Restock ${productsToRestockCount} Product${
-           productsToRestockCount > 1 ? "s" : ""
-          }`}
-      </Button>
-     </Box>
-    </Box>
-   </Modal>
-
-   {/* View Product Modal */}
-   <Modal
-    open={viewProductModal}
-    onClose={handleCloseViewProduct}
-    aria-labelledby="product-details-modal"
-   >
-    <Box
-     sx={{
-      position: "absolute",
-      top: "50%",
-      left: "50%",
-      transform: "translate(-50%, -50%)",
-      width: { xs: "90%", sm: 600 },
-      maxHeight: "90vh",
-      bgcolor: "background.paper",
-      borderRadius: 2,
-      boxShadow: 24,
-      p: 4,
-      overflow: "auto",
-     }}
-    >
-     {selectedViewProduct && (
-      <>
-       <Typography
-        id="product-details-modal"
-        variant="h5"
-        component="h2"
-        gutterBottom
-       >
-        Product Details
-        {featuredProducts.has(selectedViewProduct.productId) && (
-         <Star sx={{ ml: 1, color: "gold", verticalAlign: "middle" }} />
-        )}
-       </Typography>
-
-       <Paper elevation={0} sx={{ mt: 2, p: 2, bgcolor: "background.default" }}>
-        {/* Replace Grid with Box layout */}
-        <Box sx={{ display: 'flex', flexDirection: { xs: 'column', md: 'row' }, gap: 3 }}>
-          {/* Product Image */}
-          <Box
-            sx={{
-              flex: { md: '0 0 40%' },
-              display: 'flex',
-              justifyContent: 'center',
-              alignItems: 'flex-start',
-              m: 2,
-            }}
+  return (
+    <Box m="1.5rem 2.5rem">
+      <Box display="flex" flexDirection="column" gap={2}>
+        <Box className="flex justify-between items-center gap-4">
+          <TextField
+            label="Search Inventory"
+            variant="outlined"
+            value={searchQuery}
+            onChange={(e) => setSearchQuery(e.target.value)}
+          />
+          <Button
+            color="primary"
+            variant="contained"
+            onClick={handleOpenRestockModal}
+            disabled={selectedProductIds.length === 0}
           >
-            <Image
-              src={
-                selectedViewProduct.image
-                ? `http://localhost:8000${selectedViewProduct.image}`
-                : "/assets/default-image.png"
-              }
-              alt={selectedViewProduct.name}
-              width={200}
-              height={200}
-              style={{
-                objectFit: "contain",
-                borderRadius: "8px",
-                maxWidth: "100%",
-                height: "auto",
-              }}
-              onError={(e) => {
-                (e.target as HTMLImageElement).src = "/assets/default-image.png";
-              }}
-            />
-          </Box>
-
-          {/* Product Details */}
-          <Box sx={{ flex: '1 1 auto' }}>
-            <Typography
-              variant="h6"
-              component="h3"
-              gutterBottom
-              fontWeight="bold"
-            >
-              {selectedViewProduct.name}
-            </Typography>
-
-            <Box sx={{ mt: 2, mb: 3 }}>
-              <MuiRating
-                value={selectedViewProduct.rating || 0}
-                precision={0.5}
-                readOnly
-              />
-              <Typography variant="body2" color="text.secondary" sx={{ mt: 0.5 }}>
-                Rating: {selectedViewProduct.rating || 0} / 5
-              </Typography>
-            </Box>
-
-            <Typography
-              variant="h6"
-              color="primary"
-              fontWeight="medium"
-              sx={{ mb: 2 }}
-            >
-              ${selectedViewProduct.price.toFixed(2)}
-            </Typography>
-
-            {/* Category and Stock Side by Side */}
-            <Box sx={{ display: 'flex', flexWrap: 'wrap', mb: 2, gap: 2 }}>
-              <Box sx={{ flex: '1 1 40%', minWidth: '120px' }}>
-                <Typography variant="body2" color="text.secondary">
-                  Category
-                </Typography>
-                <Typography variant="body1">
-                  {selectedViewProduct.category || "Uncategorized"}
-                </Typography>
-              </Box>
-              <Box sx={{ flex: '1 1 40%', minWidth: '120px' }}>
-                <Typography variant="body2" color="text.secondary">
-                  Stock Quantity
-                </Typography>
-                <Box sx={{ display: "flex", alignItems: "center" }}>
-                  <Typography
-                    variant="body1"
-                    sx={{
-                      color:
-                        selectedViewProduct.stockQuantity >= 100
-                          ? "success.main"
-                          : "error.main",
-                      fontWeight: "medium",
-                    }}
-                  >
-                    {selectedViewProduct.stockQuantity}
-                  </Typography>
-                  <Chip
-                    label={
-                      selectedViewProduct.stockQuantity >= 100
-                        ? "In Stock"
-                        : "Low Stock"
-                    }
-                    color={
-                      selectedViewProduct.stockQuantity >= 100 ? "success" : "error"
-                    }
-                    size="small"
-                    sx={{ ml: 1 }}
-                  />
-                </Box>
-              </Box>
-            </Box>
-
-            <Typography variant="body2" color="text.secondary" sx={{ mt: 2 }}>
-              Product ID
-            </Typography>
-            <Typography variant="body2" sx={{ fontFamily: "monospace" }}>
-              {selectedViewProduct.productId}
-            </Typography>
-
-            {/* Featured toggle in product view */}
-            <Box sx={{ mt: 3, display: "flex", alignItems: "center" }}>
-              <Typography variant="body2" color="text.secondary">
-                Featured Status:
-              </Typography>
-              <Button
-                startIcon={
-                  featuredProducts.has(selectedViewProduct.productId) ? (
-                    <Star sx={{ color: "gold" }} />
-                  ) : (
-                    <StarBorder />
-                  )
-                }
-                variant="outlined"
-                size="small"
-                onClick={() => toggleFeatured(selectedViewProduct.productId)}
-                sx={{ ml: 2 }}
-              >
-                {featuredProducts.has(selectedViewProduct.productId)
-                  ? "Remove Featured"
-                  : "Mark as Featured"}
-              </Button>
-            </Box>
-          </Box>
+            Restock <span className="hidden md:block"> Selected Products</span>
+          </Button>
         </Box>
-       </Paper>
 
-       <Box sx={{ mt: 3, display: "flex", justifyContent: "flex-end" }}>
-        <Button onClick={handleCloseViewProduct} variant="contained">
-         Close
-        </Button>
-       </Box>
-      </>
-     )}
+        {/* Featured filter toggle */}
+        <Box display="flex" className="flex justify-between gap-3 items-center">
+          <ToggleButtonGroup
+            className="flex justify-between gap-3 items-center"
+            value={filterMode}
+            exclusive
+            onChange={handleFilterChange}
+            aria-label="product filter"
+            size="small"
+          >
+            <ToggleButton value="all" aria-label="all products">
+              <span className="text-[10px] font-medium md:text-base">All Products </span> 
+            </ToggleButton>
+            <ToggleButton value="featured" aria-label="featured products">
+              <span className="text-[10px] font-medium md:text-base">Featured ({featuredCount}) </span>  
+            </ToggleButton>
+            <ToggleButton value="regular" aria-label="regular products">
+              <span className="text-[10px] font-medium md:text-base">Regular </span> 
+            </ToggleButton>
+          </ToggleButtonGroup>
+
+          <Typography variant="body2" color="text.secondary">
+            <span className="hidden md:block rounded-full text-[10px] bg-green-100 text-green-800 md:text-base p-3">
+              {filteredProducts.length} products found
+            </span>
+          </Typography>
+        </Box>
+      </Box>
+
+      <Box
+        mt="20px"
+        height="75vh"
+        sx={{ "& .MuiDataGrid-root": { border: "none" } }}
+      >
+        <DataGrid
+          loading={isLoading}
+          rows={filteredProducts}
+          columns={columns}
+          getRowId={(row) => row.productId}
+          checkboxSelection
+          onRowSelectionModelChange={(newSelection) => {
+            setSelectedProductIds(newSelection as string[]);
+          }}
+          rowSelectionModel={selectedProductIds}
+          slots={{ toolbar: GridToolbar }}
+          sx={dataGridStyles}
+        />
+      </Box>
+
+      {/* Restock Modal */}
+      <RestockModal 
+        open={openRestockModal}
+        onClose={handleCloseRestockModal}
+        selectedProducts={selectedProducts}
+        activeProductId={activeProductId}
+        handleSetActiveProduct={handleSetActiveProduct}
+        incrementQuantity={incrementQuantity}
+        decrementQuantity={decrementQuantity}
+        handleQuantityChange={handleQuantityChange}
+        productsToRestockCount={productsToRestockCount}
+        hasProductsToRestock={hasProductsToRestock}
+        handleRestockProducts={handleRestockProducts}
+      />
+
+      {/* View Product Modal */}
+      <ProductDetailModal
+        open={viewProductModal}
+        onClose={handleCloseViewProduct}
+        product={selectedViewProduct}
+        featuredProducts={featuredProducts}
+        toggleFeatured={toggleFeatured}
+        getProductUnit={getProductUnit}
+      />
     </Box>
-   </Modal>
-  </Box>
- );
+  );
 };
 
 export default Inventory;
